Create extract_from_memory_to_folder to uncompress files & refactor extract & extract -> display_extract
